projects/bento2/requirements.txt
marimo>=0.9.0
polars-lts-cpu>=0.20.0
altair>=5.2.0
pyarrow>=12.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
optuna>=3.0.0
jpholiday>=0.1.0
